requirements.txt
Django==2.0.6
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
gevent==1.2.2
gunicorn==19.7.1
model-mommy==1.3.2
psycopg2==2.7.3
PyJWT==1.5.2
pytest==3.2.1
pytest-django==3.1.2
python-dotenv==0.6.3
raven==6.6.0
requests==2.18.3
Faker==0.8.7
redis==2.10.5
cancan==0.1.2
django-filter==1.1.0
